move body component to index
