+++ components/index.js
@@ -1,3 +1,2 @@
-export { default as Body } from './body'
 export { default as Header } from './header'
 export { default as Footer } from './footer'
